Revert "Merge branch 'dev_sprint_25_2' into feature/RDK-56052"
This reverts commit 4ed10516d858c3c75159b49a30f0b8d1188f7caa, reversing
changes made to af7dce69a4e9c1b5241c8ad4c40cf31211e53511.

+++ admanager_mpd.h
@@ -533,7 +533,7 @@ public:
 	 * @param[in] p2AdDataduration Duration of the ad break
 	 * @param[in] periodDelta Period delta
 	 */
-	void setAdMarkers(uint64_t p2AdDataduration, int64_t periodDelta);
+	void setAdMarkers(uint64_t p2AdDataduration, double periodDelta);
 
 	/**
 	 * @fn FulfillAdLoop
